requirements.txt
asyncio
opencv-python
websockets
numpy
scipy
pynmea2
python-periphery
luma.core
luma.oled
Pillow
PyYAML
